Retain fix
Currently when a user sets the fallbackDataSource, that dataSource gets put into a strongly referenced array to be used for the internal message forwarding.    The issue with this is that the user is also intented to keep a hold of the dataSource until it's time to trash.    This often times will lead to a retain cycle.

The fix is to not put the fallbackDataSource in the array of responders, and add explicit checks in the responder forwarding methods.